presque fini djikstra, marche pas encore

+++ be-graphes-model/src/main/java/org/insa/graphs/model/Label.java
@@ -3,11 +3,11 @@ package org.insa.graphs.model;
 public class Label {
     Node sommetCourant;
     boolean marque;
-    int coutRealise;
+    float coutRealise;
     Arc pere;
 
 
-    public Label(Node sommetCourant, boolean marque, int coutRealise, Arc pere){
+    public Label(Node sommetCourant, boolean marque, float coutRealise, Arc pere){
         this.sommetCourant=sommetCourant;
         this.marque=marque;
         this.coutRealise=coutRealise;
@@ -17,11 +17,11 @@ public class Label {
 //Getters
     public Node getSommetCourant(){return this.sommetCourant;}
     public boolean getMarque(){return this.marque;}
-    public int getCoutRealise(){return this.coutRealise;}
+    public float getCoutRealise(){return this.coutRealise;}
     public Arc getPere(){return this.pere;}
 
 //Setters
     public void setMarque(boolean marque){this.marque=marque;}
-    public void setCoutRealise(int coutRealise){this.coutRealise=coutRealise;}
+    public void setCoutRealise(float coutRealise){this.coutRealise=coutRealise;}
     public void setPere(Arc pere){this.pere=pere;}
 }
